requirements.txt
Flask>=3.0.0
google-api-python-client>=2.130.0
pandas>=2.2.2
python-dotenv>=1.0.1
plotly>=5.22.0
